# docs/openapi.yaml
openapi: 3.0.3
info:
  title: F4ALL Sports Talent Assessment API
  description: >
    API contract for the F4ALL AI-powered sports talent assessment platform.
    The mobile application receives a provisional on-device score, while the
    backend performs server-side verification and stores the final result.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/tests/submit:
    post:
      tags:
        - Tests
      summary: Submit a completed test
      description: >
        Submit an athlete's completed test with the provisional on-device
        result. The server will independently verify the submission.

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitTestRequest'

      responses:
        '201':
          description: Test submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitTestResponse'

        '400':
          description: Invalid test submission

        '401':
          description: Authentication required


  # ============================================================
  # VIDEO UPLOAD (chunked + resumable)
  # ============================================================
  #
  # Added in Sprint 4. The original contract referenced a video_id on
  # /api/tests/submit but defined no way to produce one.
  #
  # The flow is deliberately resumable. Athletes upload over 2G/3G links that
  # drop mid-transfer, and restarting a 6MB video from byte zero on every drop
  # burns mobile data they paid for and may not be able to replace.
  #
  #   init  ->  (status)  ->  chunk x N  ->  complete
  #
  # The client asks the SERVER which chunks arrived rather than trusting its own
  # record of what it sent, so an upload survives the app being killed.
...
  /api/results/{result_id}:
    get:
      tags:
        - Results
      summary: Get test result
      description: >
        Get the current status and scores of a submitted test.

      security:
        - bearerAuth: []

      parameters:
        - name: result_id
          in: path
          required: true
          description: Unique test result ID
          schema:
            type: string
            format: uuid

      responses:
        '200':
          description: Test result retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'

        '401':
          description: Authentication required

        '404':
          description: Result not found


  # ============================================================
  # DASHBOARD REVIEW
  # ============================================================
...
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Exchange a refresh token for a new token pair
      description: >
        Rotates the refresh token. Presenting an already-rotated token revokes
        every session for that subject, since two parties must be holding it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TokenPair"}
        "401": {description: Unknown, expired, revoked or reused refresh token}
...
  /api/dashboard/auth/login:
    post:
      tags: [Dashboard]
      summary: Official sign-in with email and password
      description: Five failed attempts lock the account for 15 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string}
                password: {type: string}
      responses:
        "200": {description: Token pair and official profile}
        "401": {description: Incorrect email or password, identical for unknown accounts}
        "423": {description: Account temporarily locked}
  /api/dashboard/auth/refresh:
    post:
      tags: [Dashboard]
      summary: Rotate an official refresh token
      description: Athlete refresh tokens are refused.
      responses:
        "200": {description: New token pair}
        "401": {description: Invalid token or inactive account}
  /api/dashboard/auth/me:
    get:
      tags: [Dashboard]
      summary: The signed-in official
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Official profile}
...
components:
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


  # ------------------------------------------------------------
  # REQUEST SCHEMAS
  # ------------------------------------------------------------
  schemas:
...
    SubmitTestRequest:
      type: object
      required:
        - test_id
        - provisional_score
      properties:
        test_id:
          type: string
          format: uuid
          description: ID of the test being submitted
          example: '550e8400-e29b-41d4-a716-446655440000'

        provisional_score:
          type: number
          format: double
          description: Provisional score calculated on the athlete's device
          example: 28

        video_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the uploaded test video
          example: '660e8400-e29b-41d4-a716-446655440000'
...
    SubmitTestResponse:
      type: object
      properties:
        result_id:
          type: string
          format: uuid
          description: Created test result ID
          example: '770e8400-e29b-41d4-a716-446655440000'

        status:
          type: string
          enum:
            - pending_sync
            - processing
            - verified
            - flagged
            - approved
            - rejected
          example: processing
    TestResult:
      type: object
      properties:
        result_id:
          type: string
          format: uuid

        athlete_id:
          type: string
          format: uuid

        test_id:
          type: string
          format: uuid

        provisional_score:
          type: number
          format: double
          nullable: true

        server_score:
          type: number
          format: double
          nullable: true

        final_score:
          type: number
          format: double
          nullable: true

        status:
          type: string
          enum:
            - pending_sync
            - processing
            - verified
            - flagged
            - approved
            - rejected
...
    TokenPair:
      type: object
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        token_type: {type: string, example: bearer}
        expires_in: {type: integer, description: Access token lifetime in seconds}
        athlete_id: {type: string, format: uuid, nullable: true}
        registered: {type: boolean}
...